feat(cross-app-connect): own the session-expired login screen + remember user
Three issues with the "session expired" branch on the transact popup:

1. Tapping "Continue" opened Privy's modal with their own picker — we
   wanted to drive sign-in entirely through the kit's matching UI.
2. The screen was a single anonymous "Continue" button — no indication
   of who the user was previously logged in as.
3. Nothing pre-highlighted the provider the user actually used last, so
   every re-login was a cold pick.

Fixes:

- **Extracted `SignInPanel`** (plus `ProviderRow` / `PhoneRow` /
  `FarcasterRow` / `RecentDot`) out of `connect/ConnectClient.tsx` into
  `cross-app/_components/SignInPanel.tsx`. ConnectClient now imports
  it; TransactClient renders it in the `!authenticated` branch with
  `onCancel={onReject}` so closing the picker rejects the cross-app
  request cleanly.
- **No more Privy modal**: dropped `useLogin()` from TransactClient.
  The headless `useLoginWithOAuth` / `useLoginWithSms` inside SignInPanel
  drive everything through our own UI.
- **Last-identity stash** at `cross-app/_lib/lastIdentity.ts`. Whenever
  `usePrivy().user` is non-null (on both connect and transact), we
  persist `{ label, provider }` to localStorage. `labelFromPrivyUser`
  derives a display string from `user.email.address`, then phone, then
  linked-social email, then the DID prefix.
- **"Welcome back, {identity}"** title on the transact session-expired
  screen when a stashed identity is available. The previously-used
  provider is forwarded to `SignInPanel` via the new `presetRecent`
  prop so its dot lights up the right row.
- **`transact.title.welcomeBack`** key added in all 17 locales.

Note re Privy session duration: it's a dashboard setting on the Privy
app (Authentication → Session), not a code option. Default is ~30 days
sliding. To raise it, edit the dashboard for the VECHAIN_PRIVY_APP_ID
app — no code change possible from the kit side.
