+++ cross-app-connect/src/app/cross-app/connect/ConnectClient.tsx
@@ -8,11 +8,8 @@ import {
     useState,
 } from 'react';
 import { useTranslation } from 'react-i18next';
-import { SiFarcaster } from 'react-icons/si';
-import { LuPhone } from 'react-icons/lu';
 import {
     useLoginWithOAuth,
-    useLoginWithSms,
     useLogout,
     usePrivy,
     useWallets,
@@ -20,36 +17,25 @@ import {
 import { useCrossAppClient } from '../_lib/client';
 import { lookupAppByUrl } from '../_lib/app-hub';
 import { getRecentProvider, setRecentProvider } from '../_lib/recent';
+import { labelFromPrivyUser, setLastIdentity } from '../_lib/lastIdentity';
 import {
     getSmartAccountAddress,
     type SmartAccountInfo,
 } from '../_lib/thor';
 import { VechainHeader } from '../../components/VechainHeader';
 import { IdentityRow } from '../../components/IdentityRow';
+import { OAUTH_PROVIDERS, type OAuthProvider } from '../../components/socials';
 import {
-    BRAND_GLYPH_COLOR,
-    FARCASTER_GLYPH_COLOR,
-    OAUTH_PROVIDERS,
-    PHONE_GLYPH_COLOR,
-    type OAuthProvider,
-} from '../../components/socials';
-import { PinInput } from './PinInput';
+    SignInPanel,
+    isIntent,
+    isOAuthIntent,
+    type IntentMethod,
+} from '../_components/SignInPanel';
 import styles from './connect.module.css';
 
 type ConnectionRequest = ReturnType<
     ReturnType<typeof useCrossAppClient>['getConnectionRequestFromUrlParams']
 >;
-
-const INTENT_METHODS = [
-    ...OAUTH_PROVIDERS.map((p) => p.id),
-    'phone',
-    'farcaster',
-] as const;
-type IntentMethod = (typeof INTENT_METHODS)[number];
-
-function isIntent(value: string | null): value is IntentMethod {
-    return !!value && (INTENT_METHODS as readonly string[]).includes(value);
-}
 
 const OAUTH_ATTEMPTED_STORAGE_KEY = 'vk-cross-app-connect:oauth-attempted';
 
@@ -74,10 +60,6 @@ function hasLinkedProvider(
     return Boolean((user as unknown as Record<string, unknown>)[intent]);
 }
 
-function isOAuthIntent(value: IntentMethod | null): value is OAuthProvider {
-    return !!value && OAUTH_PROVIDERS.some((p) => p.id === value);
-}
-
 export function ConnectClient() {
     const { t } = useTranslation();
     const client = useCrossAppClient();
@@ -95,6 +77,17 @@ export function ConnectClient() {
     const [smartAccount, setSmartAccount] = useState<SmartAccountInfo | null>(
         null,
     );
+
+    // Persist a friendly identity label + provider hint whenever a user is
+    // active. Read back on the transact "session expired" screen so we can
+    // greet the user by name and pre-highlight the right provider on
+    // re-login, instead of throwing them at a blank picker.
+    useEffect(() => {
+        if (!user) return;
+        const label = labelFromPrivyUser(user);
+        if (!label) return;
+        setLastIdentity({ label, provider: getRecentProvider() ?? undefined });
+    }, [user]);
 
     useEffect(() => {
         const hasRequesterKey =
@@ -368,348 +361,6 @@ export function ConnectClient() {
                 </div>
             </div>
         </>
-    );
-}
-
-type PanelView = 'picker' | 'phone' | 'farcaster';
-
-function SignInPanel({
-    intent,
-    onCancel,
-}: {
-    intent: IntentMethod | null;
-    onCancel: () => void;
-}) {
-    const { t } = useTranslation();
-    const [error, setError] = useState<string | null>(null);
-    const { initOAuth, loading: oauthLoading } = useLoginWithOAuth({
-        onError: (e) => setError(String(e)),
-    });
-    const { state: smsState, sendCode, loginWithCode } = useLoginWithSms();
-
-    const [view, setView] = useState<PanelView>(() =>
-        intent === 'phone'
-            ? 'phone'
-            : intent === 'farcaster'
-            ? 'farcaster'
-            : 'picker',
-    );
-    const [showOther, setShowOther] = useState<boolean>(false);
-    const [phone, setPhone] = useState('');
-    const [code, setCode] = useState('');
-    const [recent, setRecent] = useState<string | null>(null);
-
-    useEffect(() => {
-        setRecent(getRecentProvider());
-    }, []);
-
-    const rows = useMemo(() => {
-        const primary = OAUTH_PROVIDERS.filter((p) => p.tier === 'primary');
-        const other = OAUTH_PROVIDERS.filter((p) => p.tier === 'other');
-        return { primary, other };
-    }, []);
-
-    const onOAuth = (provider: OAuthProvider) => {
-        setError(null);
-        setRecentProvider(provider);
-        initOAuth({ provider }).catch((e) => setError(String(e)));
-    };
-
-    const onSendCode = async () => {
-        setError(null);
-        try {
-            await sendCode({ phoneNumber: phone });
-            setRecentProvider('phone');
-        } catch (e) {
-            setError(e instanceof Error ? e.message : t('connect.error.failedToSendCode'));
-        }
-    };
-
-    const onSubmitCode = async () => {
-        setError(null);
-        try {
-            await loginWithCode({ code });
-        } catch (e) {
-            setError(e instanceof Error ? e.message : t('connect.error.failedToVerifyCode'));
-        }
-    };
-
-    const awaitingCode = smsState.status === 'awaiting-code-input';
-    const sendingCode = smsState.status === 'sending-code';
-    const submittingCode = smsState.status === 'submitting-code';
-
-    const isRecent = (id: string) => recent === id;
-
-    return (
-        <div className={styles.card}>
-            <div className={styles.cardBodyTight}>
-                {error && (
-                    <div className={`${styles.alert} ${styles.alertError}`}>
-                        {error}
-                    </div>
-                )}
-
-                {view === 'picker' && (
-                    <div className={styles.cardBodyTight}>
-                        {rows.primary.map((p) => (
-                            <ProviderRow
-                                key={p.id}
-                                provider={p}
-                                onClick={() => onOAuth(p.id)}
-                                isDisabled={oauthLoading}
-                                isRecent={isRecent(p.id)}
-                            />
-                        ))}
-                        <PhoneRow
-                            onClick={() => setView('phone')}
-                            isRecent={isRecent('phone')}
-                        />
-                        {!showOther && rows.other.length > 0 && (
-                            <div className={styles.linkCenter}>
-                                <button
-                                    type="button"
-                                    className={styles.linkBtn}
-                                    onClick={() => setShowOther(true)}
-                                >
-                                    {t('connect.provider.moreOptions', {
-                                        count: rows.other.length + 1,
-                                    })}
-                                </button>
-                            </div>
-                        )}
-                        {showOther && (
-                            <div className={styles.cardBodyTight}>
-                                {rows.other
-                                    .filter(
-                                        (p) =>
-                                            p.id === 'discord' ||
-                                            p.id === 'github' ||
-                                            p.id === 'tiktok',
-                                    )
-                                    .map((p) => (
-                                        <ProviderRow
-                                            key={p.id}
-                                            provider={p}
-                                            onClick={() => onOAuth(p.id)}
-                                            isDisabled={oauthLoading}
-                                            isRecent={isRecent(p.id)}
-                                        />
-                                    ))}
-                                <FarcasterRow
-                                    onClick={() => setView('farcaster')}
-                                    isRecent={isRecent('farcaster')}
-                                />
-                                {rows.other
-                                    .filter((p) => p.id === 'line')
-                                    .map((p) => (
-                                        <ProviderRow
-                                            key={p.id}
-                                            provider={p}
-                                            onClick={() => onOAuth(p.id)}
-                                            isDisabled={oauthLoading}
-                                            isRecent={isRecent(p.id)}
-                                        />
-                                    ))}
-                            </div>
-                        )}
-                    </div>
-                )}
-
-                {view === 'phone' && !awaitingCode && !submittingCode && (
-                    <div className={styles.cardBodyTight}>
-                        <input
-                            type="tel"
-                            placeholder={t('connect.phone.placeholder')}
-                            value={phone}
-                            onChange={(e) => setPhone(e.target.value)}
-                            autoFocus
-                            className={styles.inputRow}
-                        />
-                        <p className={styles.muted}>
-                            {t('connect.phone.codeHint')}
-                        </p>
-                        <button
-                            type="button"
-                            className={styles.btnBrand}
-                            onClick={onSendCode}
-                            disabled={!phone || sendingCode}
-                        >
-                            {sendingCode
-                                ? t('connect.phone.sending')
-                                : t('connect.phone.sendCode')}
-                        </button>
-                        <button
-                            type="button"
-                            className={`${styles.btnGhost} ${styles.btnSm}`}
-                            onClick={() => setView('picker')}
-                        >
-                            {t('common.button.back')}
-                        </button>
-                    </div>
-                )}
-
-                {view === 'phone' && (awaitingCode || submittingCode) && (
-                    <div className={styles.cardBodyTight}>
-                        <p className={styles.mutedBody}>
-                            {t('connect.phone.codeSent', { phone })}
-                        </p>
-                        <div className={styles.pinRow}>
-                            <PinInput
-                                value={code}
-                                onChange={setCode}
-                                onComplete={(v) => {
-                                    setCode(v);
-                                    loginWithCode({ code: v }).catch((e) =>
-                                        setError(String(e)),
-                                    );
-                                }}
-                            />
-                        </div>
-                        <button
-                            type="button"
-                            className={styles.btnBrand}
-                            onClick={onSubmitCode}
-                            disabled={code.length !== 6 || submittingCode}
-                        >
-                            {submittingCode
-                                ? t('connect.phone.verifying')
-                                : t('connect.phone.verify')}
-                        </button>
-                        <button
-                            type="button"
-                            className={`${styles.btnGhost} ${styles.btnSm}`}
-                            onClick={() => {
-                                setCode('');
-                                setView('picker');
-                            }}
-                        >
-                            {t('common.button.back')}
-                        </button>
-                    </div>
-                )}
-
-                {view === 'farcaster' && (
-                    <div className={styles.cardBodyTight}>
-                        <p className={styles.mutedBody}>
-                            {t('connect.farcaster.comingSoon')}
-                        </p>
-                        <button
-                            type="button"
-                            className={`${styles.btnGhost} ${styles.btnSm}`}
-                            onClick={() => setView('picker')}
-                        >
-                            {t('common.button.back')}
-                        </button>
-                    </div>
-                )}
-
-                <button
-                    type="button"
-                    className={styles.btnGhost}
-                    onClick={onCancel}
-                >
-                    {t('common.button.cancel')}
-                </button>
-            </div>
-        </div>
-    );
-}
-
-function ProviderRow({
-    provider,
-    isRecent,
-    onClick,
-    isDisabled,
-}: {
-    provider: (typeof OAUTH_PROVIDERS)[number];
-    isRecent?: boolean;
-    onClick: () => void;
-    isDisabled?: boolean;
-}) {
-    const { t } = useTranslation();
-    const brandColor = BRAND_GLYPH_COLOR[provider.id];
-    const monoFlip =
-        provider.id === 'apple' ||
-        provider.id === 'github' ||
-        provider.id === 'twitter';
-    const iconColor = brandColor
-        ? brandColor
-        : monoFlip
-        ? 'var(--text-strong)'
-        : undefined;
-    const Icon = provider.Icon;
-    return (
-        <button
-            type="button"
-            className={styles.btnRow}
-            onClick={onClick}
-            disabled={isDisabled}
-        >
-            <Icon className={styles.rowIcon} style={{ color: iconColor }} />
-            <span className={styles.rowLabel}>
-                {t('connect.provider.continueWith', {
-                    provider: provider.label,
-                })}
-            </span>
-            {isRecent && <RecentDot />}
-        </button>
-    );
-}
-
-function PhoneRow({
-    onClick,
-    isRecent,
-}: {
-    onClick: () => void;
-    isRecent?: boolean;
-}) {
-    const { t } = useTranslation();
-    return (
-        <button type="button" className={styles.btnRow} onClick={onClick}>
-            <LuPhone
-                className={styles.rowIcon}
-                style={{ color: PHONE_GLYPH_COLOR }}
-            />
-            <span className={styles.rowLabel}>
-                {t('connect.provider.continueWithPhone')}
-            </span>
-            {isRecent && <RecentDot />}
-        </button>
-    );
-}
-
-function FarcasterRow({
-    onClick,
-    isRecent,
-}: {
-    onClick: () => void;
-    isRecent?: boolean;
-}) {
-    const { t } = useTranslation();
-    return (
-        <button type="button" className={styles.btnRow} onClick={onClick}>
-            <SiFarcaster
-                className={styles.rowIcon}
-                style={{ color: FARCASTER_GLYPH_COLOR }}
-            />
-            <span className={styles.rowLabel}>
-                {t('connect.provider.continueWithFarcaster')}
-            </span>
-            {isRecent && <RecentDot />}
-        </button>
-    );
-}
-
-function RecentDot() {
-    const { t } = useTranslation();
-    const label = t('connect.provider.lastUsed');
-    return (
-        <span
-            className={styles.recentDot}
-            role="img"
-            aria-label={label}
-            title={label}
-        />
     );
 }
