+++ cross-app-connect/src/app/i18n/locales/ko.json
@@ -126,7 +126,8 @@
             "signIn": "계속하려면 로그인",
             "couldNotLoad": "요청을 불러올 수 없습니다",
             "signMessage": "메시지 서명",
-            "signData": "데이터 서명"
+            "signData": "데이터 서명",
+            "welcomeBack": "다시 만나서 반갑습니다, {{identity}}"
         },
         "subtitle": {
             "sendVet": "{{recipient}} 에게 {{amount}} VET",
